Money System Initial Version
Really big issues with this right now:
The system has to check and do the math each and every time you want to check what costs are.  

To improve it, I need to separate the checks for each resource ,and the check to adapt the island's payouts, then i need to only update when the total resources will change (player actions, ships creation or deletion, island updates)

The system currently tallies all existing resources to find the value of the four, and will add or remove 1/2 price depending on how much of something an island has.  It's now a good indicator of economic health.  A balanced economy is hard to make money from, but an unbalanced one incentives the player to trade to balance it.

Function is at the bottom of game.c, returns a struct containing all the variables for the costs, and there is a comment documenting how the params work.

Removed some app logs and other comments.