Restructured Menu Drawing, player_update split
Menu drawing now works as seen in the comment above the menu drawing function.  Updating the player has been split into a few functions, and I plan on making a) the finding of if a button is pressed easier by going to a regular system instead of the stupid uphit downhit one.  Using a function for now until I decide to do that.

Adding in structure here to allow to add in the upgrades feature.  More constants are being put in to make the code more readable.