Finished Menu Updates
Menu is now fully ready to be worked on again, to add in new functions.  Even managed to get back one percent of ram by making it into functions.

Odd issue with some freezing that may or may not be consistent or based on something that went away after a second.