Added Navigation Compass
It works using the idea of similar triangles.  I didn't comment on the code well at all.  Future me, if you see this, I drew a paper about how it all works, so try to decipher and comment on it later.  For now, it works, and honestly a lot of it was trial and error rather than actual understanding.