Moved Islands, Changed icons, added islands
Icons now show what type of islands they are, so the player can go out in search of metal islands, or such.

I moved the islands apart a bit now that you can see where other islands are.

Also added the ability to start and stop with the middle button on the watch.  This also stops the "player-loop" where the player runs into something over and over again.

This does happen still with ships.