Island faction system.  Other misc changes
Got the basics working for this.  Islands now hate or like you based on a number of factors, with islands moving towards the opposite faction of any island that has a resource it wants, and with player actions changing who islands ally with.  Islands no longer trade with those who pillage all their ships over and over.

Fixed more things with the persistent ships, there is still more to polish out, but things are stable enough for a commit.