Final fixes to new system
Changed a few things in the menus to prevent accidental button pressing.

Updated showing the resource of ships to show the actual type, and none if the ship cargo hold is empty

Adjusted ship battles to account for variable ship cargo

Allowed viewing island allegiance in menus, even if the island does not want to trade with you

Made minimap cover a larger range to ensure you can see icons for all the islands allegiances.

Forgot how to spell allegiance.