All menus functional, needs more work
Menus now all will scroll properly, and you can buy and sell ships.

The proper data isn't saved, so the game slowly died as you open and close it, this is not stable.

The menus also don't appear in a proper order yet, any menu will cause the others not to appear, causing menus to "float" to the right edge of the screen.