Ships now buy and sell, island has a wallet
None of the stuff is saved yet, but that whole cleanup process is what comes next.  After this, I plan on renaming some menus, setting values to be saved to persistent memory, and bugchecking, and this will likely be the final release version.  Going to move onto a new game, that I think will be interesting.